Address review: share FakeVolume fixture, parametrize, generalize merge
- Move FakeVolume + fake_volume fixture into tests/conftest.py so they're
  shared/importable rather than living in one test module.
- Parametrize the TrainResult save-survives test across all Frameworks; drop
  the PR-description-style docstrings for present-tense behavior descriptions.
- _merge_recipe no longer hardcodes SlimeRecipe as the MRO stop point: it
  breaks at the framework base recipe (immediate BaseTrainRecipe subclass), so
  the declared-fields logic applies to any framework's recipe subclasses.
