+++ tests/test_metadata_persistence.py
@@ -1,16 +1,13 @@
-"""Save-path regression tests for the metadata layer.
+"""Drive the real ``save()`` chain to completion without Modal or a GPU.
 
-Two bugs reached `main` because nothing exercised a run all the way to
-``save()`` without a full GPU train: a ``Framework`` enum that wasn't
-JSON-serializable, and ``Volume.reload()`` crashing when the volume isn't
-attached (local driver / a function that doesn't mount it). These tests drive
-the real ``save()`` chain against an in-memory volume — no Modal, no GPU — so
-that path is covered in CI.
+``TrainingRun.save()`` and ``TrainResult.save()`` are exercised against an
+in-memory ``FakeVolume`` (see ``conftest.py``) so the full serialize-and-write
+path is covered in CI: the payload must stay JSON-serializable, and ``save()``
+must complete even when ``Volume.reload()`` is unavailable.
 """
 
 from __future__ import annotations
 
-import io
 import json
 import os
 
@@ -19,60 +16,11 @@ import pytest
 from modal_training_gym.common import run as run_mod
 from modal_training_gym.common.framework import Framework
 from modal_training_gym.common.train_result import TrainResult
-from modal_training_gym.utils import metadata
 from modal_training_gym.utils.metadata import MetadataStore
 
 
-class _FakeVolume:
-    """In-memory stand-in for a Modal Volume that is *not* attached.
-
-    ``reload()`` raises like a real unattached/local volume; reads and writes
-    operate on an in-memory dict. A correct metadata layer must still complete a
-    ``save()`` against this — reload is only a freshness hint.
-    """
-
-    def __init__(self) -> None:
-        self.files: dict[str, bytes] = {}
-
-    def reload(self) -> None:
-        raise RuntimeError("reload() can only be called from within a running function")
-
-    def read_file(self, path: str):
-        if path not in self.files:
-            raise FileNotFoundError(path)
-        return [self.files[path]]
-
-    def remove_file(self, path: str) -> None:
-        if path not in self.files:
-            raise FileNotFoundError(path)
-        del self.files[path]
-
-    def batch_upload(self):
-        files = self.files
-
-        class _Batch:
-            def __enter__(self):
-                return self
-
-            def __exit__(self, *_exc):
-                return False
-
-            def put_file(self, fileobj: io.BytesIO, path: str) -> None:
-                files[path] = fileobj.read()
-
-        return _Batch()
-
-
-@pytest.fixture
-def fake_volume(monkeypatch) -> _FakeVolume:
-    vol = _FakeVolume()
-    monkeypatch.setattr(metadata, "_metadata_volume", lambda: vol)
-    return vol
-
-
 def test_training_run_save_survives_unmounted_volume(fake_volume):
-    """The local-driver path that crashed: TrainingRun.save() must complete even
-    when reload() raises, and persist a valid JSON payload."""
+    """TrainingRun.save() completes when reload() raises, and persists valid JSON."""
     run_mod.TrainingRun(
         training_run_id="t1", framework=Framework.SLIME, config={}
     ).save()
@@ -81,19 +29,17 @@ def test_training_run_save_survives_unmounted_volume(fake_volume):
     assert json.loads(blob)["framework"] == "slime"
 
 
-def test_train_result_save_survives_unmounted_volume(fake_volume):
-    """TrainResult.save() is the path that hit both the reload crash and the
-    non-serializable Framework enum (asdict keeps the enum)."""
-    TrainResult(app_name="a", framework=Framework.MILES, training_run_id="t2").save()
+@pytest.mark.parametrize("fw", list(Framework))
+def test_train_result_save_survives_unmounted_volume(fake_volume, fw):
+    """TrainResult.save() completes when reload() raises, for every framework."""
+    TrainResult(app_name="a", framework=fw, training_run_id="t2").save()
 
     blob = fake_volume.files[f"{MetadataStore.TRAIN_RESULTS.value}/t2.json"]
-    assert json.loads(blob)["framework"] == "miles"
+    assert json.loads(blob)["framework"] == fw.value
 
 
 @pytest.mark.parametrize("fw", list(Framework))
 def test_train_result_payload_is_json_serializable(fw):
-    """Every Framework must round-trip through plain json.dumps — guards against
-    regressing the (str, Enum) base back to a bare Enum."""
     payload = TrainResult(app_name="a", framework=fw, training_run_id="t")._to_dict()
     assert json.loads(json.dumps(payload))["framework"] == fw.value
